update final semoga
